extract auth logic into own file
