+++ server/routes/dashboardRoutes.js
@@ -5,6 +5,6 @@ const router = express.Router();
 
 router.get('/getCookie', dashboardController.getCookie, (req, res) => {
   res.status(200).json(res.locals.cookie);
-})
+});
 
 module.exports = router;
